src/app/main/canvas-select/overlays/overlay-texts/shared/directives/editable-text/editable-text.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { EditableTextComponent } from './editable-text.component';

describe('Component: EditableText', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [EditableTextComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([EditableTextComponent],
      (component: EditableTextComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(EditableTextComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(EditableTextComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-editable-text></app-editable-text>
  `,
  directives: [EditableTextComponent]
})
class EditableTextComponentTestController {
}

